[jacodb-core] In JarFacade, avoid caching JarEntries returned by JarFile
Some implementations of JarEntry are inner classes referencing their JarFiles
for doing some internal checks,e.g., on getting input stream. Such caching
may require the JarFiles to always be open. To address the issue, copies
of all JarEntries should be cached.
